requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
requests
ujson
line-bot-sdk
fastapi-admin
gunicorn
psycopg2
